Update build.gradle for compatibility and dependency changes
This commit updates the  file to make the library compatible with newer versions of Android Studio and Gradle, and to use updated versions of dependencies. Specifically:

- The Kotlin plugin has been updated to version 1.6.10.
- The Android Gradle plugin has been updated to version 7.1.1.
- The repositories now include Maven local in addition to Google and Maven Central.

These changes should make it easier for developers to use this library in their projects, while maintaining compatibility with the latest Android development tools.